Move `BoxCow` and `SimpleCow` to a common module
